Added additional view controller for the confirmation page.
