+++ Stytch/Source/Strings/en.strings
@@ -2,7 +2,7 @@
 "stytch_login_description" = "Sign up or log in with your email, no password needed.";
 "stytch_login_email_hint" = "example@email.com";
 "stytch_login_button_title" = "Sign in with email";
-"stytch_login_waiting_verification_title" = "Email sent to %@.";
+"stytch_login_waiting_verification_title" = "Email sent to %@";
 "stytch_login_waiting_verification_description" = "Didn't get an email? Check your spam folder, or";
 "stytch_login_waiting_verification_button_title" = "Resend magic link";
 "stytch_watermark_powered_by" = "Powered by";
